moved item display to subcomponents
